Enable configurable mouse button 1, 2 and 3 functions in all puzzles
Now there are new default functions on pressing mouse buttons:
	* clicking mouse 1 on neighbors moves a character if possible
	* clicking mouse 2 anywhere activates undo
	* clicking mouse 3 calculates path and tries to auto move there

The puzzles may configure all this in their press_cell method.

If press_cell method returns False then the defaults described above
are activated, i.e.:
	handled = puzzle.press_cell(cell, button)
	if handled is False:
		# default button mapping if not handled
		if button == 1:
			handled = "move"
		if button == 2:
			handled = "undo"
		if button == 3:
			handled = "move-path"

If the method returns True this means (as usual) that the puzzle handled
this press, and no default button mappings will be activated.

Puzzles could return "move", "undo" or "move-path" strings directly
instead of False or True, changing the button mappings, for example:

	def press_cell(self, button):
		if button == 2:
			return "move-path"
		if button == 3:
			return "undo"
		return True

This effectivelly swaps the default button 2 and 3 functions, and
disables the default button 1 function in that puzzle.

The "move-path" function uses the solution interface to make auto moves
in the best path to the pressed cell. The AUTO_MOVE_DELAY is used that is
4 times faster than the default solution re-play (SOLUTION_MOVE_DELAY).

The "move-path" stops at the first drop (or the enemy holding the drop).
It does pass empty enemies effectivelly beating them on the way.

The move-char during solution now checks whether the move is possible;
this makes some complex puzzles to be ready to report real solutions.
For example, in SwitchBox the gate may be open when "move-path" is
started, but it becomes closed during the move, so this check is needed.

Adapt puzzles to use the new extended interface. Some puzzles partially
disable the default functions, since they want to handle mouse
differently at times. In some puzzles (like Minotaur) having move-path
would be non intuitive, so this is disabled.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>