Implement a hacky method to scale large maps to fit display
On Ctrl-S it now tries to scale to fit display. This should only be used
for informational purposes. In this mode the mouse does not report the
cell coordinates properly.

The actual dimensions of the window are not changed, it will still be
larger than the display. Only the content is scaled by copying and
overriding the screen content every tick. This may produce artifacts.

The displayed content is scaled proportionally, and the resulting size is
(DISPLAY_WIDTH, DISPLAY_HEIGHT - 68) to allow space for title/taskbar.

If the scale is not needed, Ctrl-S does nothing, but the state stays
toggled (for the next huge level).

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>