Fix fullscreen (F11) in most cases to scale to the actual content
Use pygame SCALED flag to set_mode. In most of the screens and
resolutions it should scale to the actual game content, so everything is
visible. Still allow Shift-F11 to do what it did (without SCALED flag).

This is (fullscreen using F11) currently the recommended way to see the
whole window content. Using Ctrl-S is not recommended.

Maybe in the future the window will be automatically scaled for huge
maps, but currently it is not.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>