statusmessage: Add useful function set_quick_status_message
It is often useful to notify a user about some change for a short time.

set_quick_status_message is designed for this. It gets message and sets a
status message at next-to-highest priority 1 and duration
QUICK_STATUS_MESSAGE_DURATION (that is 3 seconds by default).

The message source is unique per caller (function-name or object).

If optional flag is specified then its value (as "on" or "off") is
appended to the given message. I.e.:

	set_quick_status_message("The music is now", is_music_enabled)

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>