BarrelPuzzle: Count moves and shifts while playing, and display them
Use the last_char_move functionality to know whether the move was a
barrel shift and maintain the number of played moves and shifts.

Display the moves and shifts done since the level (room) start at the top
left corner of the puzzle map.

Currently Undo does not work. It will be fixed later.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>