test: Implement simple yet powerfull TestSuite
This is a new simple test framework, good for unit testing.

The usage is:

	test = TestSuite("MyClass")
	test.eq(value, "expected")
	test.gt(cost, 100)
	test.len(coords, 2)
	test.not_in(None, coords)
	test.not_none(solution)
	test.finalize()

All current methods:
	reset()
	ok(is_passed, error)
	eq(value1, value2)
	ne(value1, value2)
	gt(value1, value2)
	lt(value1, value2)
	is_cell(cell)
	is_in(elem, mapping)
	not_in(elem, mapping)
	is_instance(item, type0)
	not_instance(item, type0)
	has_attr(obj, name)
	no_attr(obj, name)
	len(collection, l)
	none(value)
	not_none(value)
	finalize()

By default it just writes "All 5 MyClass tests passed" or "Only 3 of 5
MyClass tests passed" and lists all failed tests.

With constructor options it is possible to raise exception with call
trace on the first failure or/and list passed tests too.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>