solution: Do not use a space delimiter in solution string
The standard Sokoban notation does not use a space delimiter, so let's do
the same. We may cut it on 70 chars somewhere later, but not in solution
module itself.

Adapt the solution items like "press()" and "goto()" to the fact there
are no space delimiters now. Use "{press()}" and "{goto()}".

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>