test: Add a new feature of automatically printing caller function name
When verbose_on_pass or on failure, print caller function name
automatically. This is always desired.

The caller function name is considered the second from the last. This may
sometimes be not what we want (when running under running layers, like
runpy.run_path(), but this is good enough in most of cases.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>