debug: Some improvements in Debug methods
Rename Debug.enable_feature to enable.

Add Debug.has(feature) that queries whether the given feature is enabled.

So it is now possible to write:
	if debug.has(DBG_SOLV):
		grid.show_map()

Also, in debug() call support more formats for msg. If it is list or
tuple of strings then just print its elements. If it is dict like
{field: value}, then print "field=value" for each field.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>