Grid: Use bitarray.search instead of itersearch
In the new bitarray releases search() now returns the iterator, and the
old itersearch() was removed. However the old search() will work too,
even if it returns list instead of iterator. So use the new API and hope
users will upgrade bitarray one day to get the intended iter behaviour.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>