Correctly implement default-map-size and screen-size-fitting-display
Now DEFAULT_MAP_SIZE (was: 13x13) is set from the actual display size.
On resolutions 1366x768 it may be 15x9. On HDD it is still 13x13.

Both Ctrl-S (fit screen content to display) and the default-map-size try
to fit the visible screen into (DISPLAY_W, DISPLAY_H - EXTRA_DISPLAY_H).
Currently EXTRA_DISPLAY_H is 71 pixels that should be all right for panel
and window title in the current GNOME and Windows 11.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>