CellActor: Improve animate to always call the previous animation callback
In animate method, check the previous animation of the actor (if any) and
execute its on_finished if any, otherwise the animation will be replaced,
and the previous callback will not be called.

This fixes some puzzles that became broken on long animation, for example
the move-path didn't call all moves in BarrelPuzzle, just the last one
(the only one with the fully finished animation). This resulted in
incorrect move count.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>