Fix crash on changes after undo that is completely empty in some cases
After move-path and then undo-move, there was a crash, since there was an
attempt to add changes without any undo_frame. Fix this by ensuring at
least one undo_frame exists when in_level.

Another fix would be to call game.start_move() to ensure the undo_frame
always exists, but do this fix it so that start_move is not needed.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>