Add new config flag CHAR_ANIMATION_SMOOTH_CUT to control animation type
Currently on fast animation (like move-path) the char cuts corners. This
is one animation type to cut animation and replace it by the new one.
Another type would be to jump into the old non-finished animation target
and continue new animation for there (and not from the unfinished
intermediate char position). When CHAR_ANIMATION_SMOOTH_CUT is True,
use the first method, otherwise the second jump method.

Maybe later there would be one more uniform-jump method, but it's
difficult to predict future animations, so it's not implemented.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>