In create_sokoban_solver allow show_map without descr
It is not always desired to write descr "Created solver for map" before
the map. For example, sokodun would better just show the map without this
header line (that adds no value). However tests/sokobansolver may still
continue to show this header for info when verbose.

Allow show_map flag in create_sokoban_solver to be False, True and descr
str, and pass the string or False in tests/sokobansolver, and just True
or False in sokodun (as before).

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>