sokobansolver: Add proper support for zsb puzzles in configure
Now grid.check_zsb is called in SokobanSolver.configure and not in
grid.prepare_sokoban_solution with proper barrel set. So that it is
properly detected in both GUI and non-GUI.

Also, do not set default solution_alg. In SokobanSolver.configure use
this logic: if solution_alg is not set, use ASTAR for zsb and with
return_first, and BFS otherwise.

Remove this logic of solution_alg for zsb from BarrelPuzzle.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>