debug: Enhance ProgressLine with 2 new features
Add new constructor param same_line=True. Setting it to False causes the
output to be regular lines with endl, and never be erased. This may be
useful when all progress output is better to be saved or shown for
debugging purposes rather than lost at the end.

Change the default for constructor param to max_len=None. If not
explicitely given, it is now auto-determined from the terminal size.

Shorten some attribute names.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>