Add and redefine SOLUTION_ALG_UCS and SOLUTION_ALG_GREED
The algorithm that was previously called Greedy is actually UCS (Uniform
Cost Search). However add actual Greedy algorithm that is similar to A*,
but more agressive in some cases, and thus not optimal.

The UCS considers only the past (position.total_nums). A* considers both
past and future cost with factor 1. Greedy considers both past and future
cost with factor (0.82, 1.22). The Greedy algorithm may be changed later.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>