colorize: Add new module colorize.py for all colorize purposes
For now it contains ansi-color constants and functions:
	colorize
	colorize_auto

Usage:
	colorize(string, COLOR_YELLOW)
	colorize_auto(debug_line)

It will turn colors off if not on terminal or when env NO_COLOR variable
is set. In this case the original string or line is returned as is.

Note that currently colorize_auto may fix strings. It may add a space
after a colon if dict is autodetected and strip some extra spaces.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>