Grid: Enhance value of show_dead in show_map to support colors
No show_dead may be False, True, or 'auto', or 'color'.

When 'color', then dead-barrel-cells are marked as red instead of "x".
But if colors are disabled, then no marking will be done.

When 'auto', then if colors are enabled they are used, otherwise the "x"
is used.

Use show_dead='auto' in sokodun on --show-dead option.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>