Grid: Disable static surrounding deadlock in try_pull
On prepare_solution for Sokoban, pull from barrels on plates is used to
determine the dead_barrel_bits. The existing surrounding deadlock
although is correct, but does not considers the initial char position
that may start from surrounding deadlock from the pull perspective.

Just disable this (the only) deadlock for pulls completely. If the char
is in this situation there will be no further pulls anyway.

Adjust the test for this pull deadlock.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>