debug: When msg is dict and value is dict, sort keys recursivelly
debug(msg) supports msg that is dict. In this case every key/value is
printed as it was a separate msg with the same other params. Do not sort
these keys, respect the user provided order. However, if the included
value is dict on itself, sort its keys recursivelly. This is usually what
is expected from the debug output, although unfortunately cells will be
sorted by X first and not by Y first, since the sorting is default.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>