sokodun: Add new option --record-file FILE to compare results to records
Now it is possible to specify the records file with cost (or pair of
push-optimal and move-optimal costs) per line, and compare the actual
results to the stored records. For each level (if the requested levels
correspond to the records file; this is the user's responsibility), the
comparision string is printed in color (if available), i.e.
"Record recorded-num_str: comparision", like:
	Record 20/6: OK  # or: WORSE, or: NEW RECORD, or: N/A

Update colorize_auto with these new capitalized strings.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>