SokobanSolver: Use Hungarian algorithm for get_min_solution_cost
Currently get_min_solution_cost is used in:
	* estimate_solution_depth - for BFS and DFS
	* super_position.solution_cost - for A* and Greedy
	* process_position - to prune position by solution lower bound

The old greedy matching with backtracking is still available as
old_get_min_solution_cost, but may be removed. Both should work.

The old complexity is O(m*n), the new one is O(m^2*n). But Hungarian
algorithm guarantees to find the best lower bound, that can potentially
lead to a faster solution for A* and Greedy.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>