sokodun: Support update of record file via env UPDATE_RECORDS
Updating record file is risky if a user specified a wrong record file, so
support it by supplying env variable UPDATE_RECORDS=1.

In this case, if together with --record-file then the script writes at
the end either "Record file <filename> is updated" (if new records
registered) or "... is not updated" if the file was unchanged.

This may be used to create a record file from scratch as well, just touch
the new records file and run:
	UPDATE_RECORDS=1 sokodun -R new-collection-record-file collection

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>