SokobanSolver: Make USC and BFS algorithms optimal
When a position is reparented (a better path to it found), then
recalculate all descendant positions down. This includes per position:
	* calc_nums
	* repush to heap if the position was (or is) already in heap

So all children of the improved position are improved too.

This fixes optimality of USC (both push-optimal and move-optimal) and BFS
(only push-optimal, since BFS has guaranteed full depth search, but it
may currently stop earlier for move-optimal search; may be fixed later).

Also, introduce per position best_key instead of best_position_keys dict
in solver, used when a position is entered into heap (A*, Greedy, UCS).
And do not unset it now after the position is poped from heap. Since the
same position can only re-enter to heap with a better key (and for UCS it
never re-enters the heap again), otherwise there is a bug somewhere.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>