SokobanSolver: Fix A* optimality for some maps due to initial position
If the solution incountered initial position again, it was created again.
This was needed for Circular Sokoban puzzles. But for regular Sokoban
puzzles this broke A* optimality for some maps.

Now the initial position is created in such way that it will not be
recreated, unless it is a Circular puzzle. In which case there will be 2
positions, one initial that is not marked as solved, and one non-initial
that will get improved solution just like a regular position with
is_solved.

In both of these 2 positions in Circular puzzle char and barrels are the
same, so persistent_id is the same. Solve this issue by appending "+" to
every solved position.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>