sokobansolver: Properly implement reverse-barrel-mode for non GUI
In create_sokoban_solver (that is used in sokodun and in tests, and not
in GUI) call new reverse_barrel_map function. The idea is the same as in
GUI - swap plates and barrels. But here all plates and barrels in the map
are swapped, and not just inside the room, as in GUI.

Update tests accordingly.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>