sokobansolver: Fix A* optimality
When a better path to a position is discovered all child edges should be
reconsidered (relaxed) and not just the current children. For A* store
all possible continuations in child_edges. The parent is still the
current best path to a position. But now all possible parents are
recorded for A* and potential children are reparented back properly.

This fixes the A* optimality. Now A* is push-optimal and move-optimal.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>