MinotaurPuzzle: Ignore button 3 for now and small tweaks
The standard button 3 meaning (move-path) can't work in MinotaurPuzzle
yet, since the minotaur move is not finalized, but divided to 2 (or
num_moves) parts, and when the solution for move-path is played the
minotaur does not do 2 moves, but only 1 that breaks solution play.

For now just ignore button 3. The wait move may be achieved by clicking
button 1 on the character.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>