Show level title and goal at the top and not at the middle
Show the level title and goal for normal levels in the first row instead
of the middle of the window. This way the text does not cover the puzzle
too much and it is possible to start to play it earlier.

For the virtual puzzles (inlcluding MainScreen), the title and goal are
still displayed in the middle.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>