BarrelPuzzle: Fix old-generation level generation without use_ng
After successful generate_random_free_path the intention is to put barrel
into its neighbour (char_cell) and not into old char_cell as it was after
the commit mentioned in Fixes.

Also, pass area to generate_random_free_path, so that no floor outside
the area is replaced.

Also, rework the whole pull_barrel_randomly to use a new style with cells
and pull terminology.

Fixes: 88246a9705c4 ("Improve generate_random_free_path")
Signed-off-by: Mikhael Goikhman <migo@freeshell.org>